intersectionをwebpackへ挿入
